Add second iteration of sample app
