+++ README.md
@@ -1,3 +1,2 @@
 # goprod
 Sample app with Go using web frameworks i'm used to and love (Guice (DI), Hibernate (Orm), Log4j (logs) etc) 
-
