+++ internal/models/product.go
@@ -1,0 +1,9 @@
+package models
+
+import "gorm.io/gorm"
+
+type Product struct {
+	gorm.Model
+	Code  string
+	Price uint
+}
